+++ src/main/java/com/criticalrange/transformer/LightingBatchTransformer.java
@@ -13,14 +13,15 @@ import org.objectweb.asm.Opcodes;
  */
 public class LightingBatchTransformer extends BaseTransformer {
 
-    private static final String TARGET_CLASS = "com/hypixel/hytale/server/core/universe/world/lighting/ChunkLightingManager";
+    private static final String TARGET_CLASS = "com.hypixel.hytale.server.core.universe.world.lighting.ChunkLightingManager";
+    private static final String TARGET_CLASS_INTERNAL = "com/hypixel/hytale/server/core/universe/world/lighting/ChunkLightingManager";
     
     public static final String ENABLED_FIELD = "$catalystLightingBatchEnabled";
     public static final String BATCH_SIZE_FIELD = "$catalystLightingBatchSize";
 
     @Override
     protected boolean shouldTransform(String className) {
-        return TARGET_CLASS.equals(className);
+        return className.equals(TARGET_CLASS);
     }
 
     @Override
@@ -81,10 +82,10 @@ public class LightingBatchTransformer extends BaseTransformer {
 
         static void initFields(MethodVisitor mv) {
             mv.visitInsn(Opcodes.ICONST_0);
-            mv.visitFieldInsn(Opcodes.PUTSTATIC, TARGET_CLASS, ENABLED_FIELD, "Z");
+            mv.visitFieldInsn(Opcodes.PUTSTATIC, TARGET_CLASS_INTERNAL, ENABLED_FIELD, "Z");
             
             mv.visitIntInsn(Opcodes.BIPUSH, 8);
-            mv.visitFieldInsn(Opcodes.PUTSTATIC, TARGET_CLASS, BATCH_SIZE_FIELD, "I");
+            mv.visitFieldInsn(Opcodes.PUTSTATIC, TARGET_CLASS_INTERNAL, BATCH_SIZE_FIELD, "I");
         }
     }